Rebase to pass CI and reflect suggestions
